// clients/typescript/models/HealthCheckFromWorker.ts
/**
 * Cloud API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class HealthCheckFromWorker {
    'environmentId': string;
    'messageId'?: string;
    'executionId': string;
    'message': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "environmentId",
            "baseName": "environment_id",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "messageId",
            "baseName": "message_id",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "executionId",
            "baseName": "execution_id",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return HealthCheckFromWorker.attributeTypeMap;
    }

    public constructor() {
    }
}